+++ core/modules/modsendrecurringinvoicebymail.class.php
@@ -18,20 +18,20 @@
  */
 
 /**
- * 	\defgroup   sendfacrecmail     Module sendfacrecmail
- *  \brief      sendfacrecmail module descriptor.
+ * 	\defgroup   sendrecurringinvoicebymail     Module sendrecurringinvoicebymail
+ *  \brief      sendrecurringinvoicebymail module descriptor.
  *
- *  \file       htdocs/sendfacrecmail/core/modules/modsendfacrecmail.class.php
- *  \ingroup    sendfacrecmail
- *  \brief      Description and activation file for module sendfacrecmail
+ *  \file       htdocs/sendrecurringinvoicebymail/core/modules/modsendrecurringinvoicebymail.class.php
+ *  \ingroup    sendrecurringinvoicebymail
+ *  \brief      Description and activation file for module sendrecurringinvoicebymail
  */
 include_once DOL_DOCUMENT_ROOT .'/core/modules/DolibarrModules.class.php';
 
 
 /**
- *  Description and activation class for module sendfacrecmail
+ *  Description and activation class for module sendrecurringinvoicebymail
  */
-class modsendfacrecmail extends DolibarrModules
+class modsendrecurringinvoicebymail extends DolibarrModules
 {
 	/**
 	 * Constructor. Define names, constants, directories, boxes, permissions
@@ -48,7 +48,7 @@ class modsendfacrecmail extends DolibarrModules
 		// Use here a free id (See in Home -> System information -> Dolibarr for list of used modules id).
 		$this->numero = 500331;		// TODO Go on page https://wiki.dolibarr.org/index.php/List_of_modules_id to reserve id number for your module
 		// Key text used to identify module (for permissions, menus, etc...)
-		$this->rights_class = 'sendfacrecmail';
+		$this->rights_class = 'sendrecurringinvoicebymail';
 
 		// Family can be 'base' (core modules),'crm','financial','hr','projects','products','ecm','technic' (transverse modules),'interface' (link with external tools),'other','...'
 		// It is used to group modules by family in module setup page
@@ -58,22 +58,22 @@ class modsendfacrecmail extends DolibarrModules
 		// Gives the possibility for the module, to provide his own family info and position of this family (Overwrite $this->family and $this->module_position. Avoid this)
 		//$this->familyinfo = array('myownfamily' => array('position' => '01', 'label' => $langs->trans("MyOwnFamily")));
 
-		// Module label (no space allowed), used if translation string 'ModulesendfacrecmailName' not found (sendfacrecmail is name of module).
+		// Module label (no space allowed), used if translation string 'ModulesendrecurringinvoicebymailName' not found (sendrecurringinvoicebymail is name of module).
 		$this->name = preg_replace('/^mod/i','',get_class($this));
-		// Module description, used if translation string 'ModulesendfacrecmailDesc' not found (sendfacrecmail is name of module).
+		// Module description, used if translation string 'ModulesendrecurringinvoicebymailDesc' not found (sendrecurringinvoicebymail is name of module).
 		$this->description = "Send generated invoice by email";
 		// Used only if file README.md and README-LL.md not found.
 		$this->descriptionlong = "This module hooks onto the recurring invoice generation to automatically send the generated PDF.";
 
 		$this->editor_name = 'Bugness';
-		$this->editor_url = 'https://www.bugness.org';
+		$this->editor_url = 'https://code.bugness.org/Dolibarr/sendrecurringinvoicebymail';
 
 		// Possible values for version are: 'development', 'experimental', 'dolibarr', 'dolibarr_deprecated' or a version string like 'x.y.z'
-		$this->version = '0.2.2';
+		$this->version = '0.2.3';
 
         //Url to the file with your last numberversion of this module
         //$this->url_last_version = 'http://www.example.com/versionmodule.txt';
-		// Key used in llx_const table to save module status enabled/disabled (where SENDFACRECMAIL is value of property name of module in uppercase)
+		// Key used in llx_const table to save module status enabled/disabled (where SENDRECURRINGINVOICEBYMAIL is value of property name of module in uppercase)
 		$this->const_name = 'MAIN_MODULE_'.strtoupper($this->name);
 		// Name of image file used for this module.
 		// If file is in theme/yourtheme/img directory under name object_pictovalue.png, use this->picto='pictovalue'
@@ -90,39 +90,39 @@ class modsendfacrecmail extends DolibarrModules
 			'tpl' => 0,                                      	// Set this to 1 if module overwrite template dir (core/tpl)
 			'barcode' => 0,                                  	// Set this to 1 if module has its own barcode directory (core/modules/barcode)
 			'models' => 0,                                   	// Set this to 1 if module has its own models directory (core/modules/xxx)
-			//'css' => array('/sendfacrecmail/css/sendfacrecmail.css.php'),	// Set this to relative path of css file if module has its own css file
-	 		//'js' => array('/sendfacrecmail/js/sendfacrecmail.js.php'),          // Set this to relative path of js file if module must load a js on all pages
+			//'css' => array('/sendrecurringinvoicebymail/css/sendrecurringinvoicebymail.css.php'),	// Set this to relative path of css file if module has its own css file
+	 		//'js' => array('/sendrecurringinvoicebymail/js/sendrecurringinvoicebymail.js.php'),          // Set this to relative path of js file if module must load a js on all pages
 			'hooks' => array('cron'), 	// Set here all hooks context managed by module. To find available hook context, make a "grep -r '>initHooks(' *" on source code. You can also set hook context 'all'
 			'moduleforexternal' => 0							// Set this to 1 if feature of module are opened to external users
 		);
 
 		// Data directories to create when module is enabled.
-		// Example: this->dirs = array("/sendfacrecmail/temp","/sendfacrecmail/subdir");
-		$this->dirs = array("/sendfacrecmail/temp");
+		// Example: this->dirs = array("/sendrecurringinvoicebymail/temp","/sendrecurringinvoicebymail/subdir");
+		$this->dirs = array("/sendrecurringinvoicebymail/temp");
 
-		// Config pages. Put here list of php page, stored into sendfacrecmail/admin directory, to use to setup module.
-		$this->config_page_url = array("setup.php@sendfacrecmail");
+		// Config pages. Put here list of php page, stored into sendrecurringinvoicebymail/admin directory, to use to setup module.
+		$this->config_page_url = array("setup.php@sendrecurringinvoicebymail");
 
 		// Dependencies
 		$this->hidden = false;			// A condition to hide module
 		$this->depends = array('modFacture');		// List of module class names as string that must be enabled if this module is enabled. Example: array('always1'=>'modModuleToEnable1','always2'=>'modModuleToEnable2', 'FR1'=>'modModuleToEnableFR'...)
 		$this->requiredby = array();	// List of module class names as string to disable if this one is disabled. Example: array('modModuleToDisable1', ...)
 		$this->conflictwith = array();	// List of module class names as string this module is in conflict with. Example: array('modModuleToDisable1', ...)
-		$this->langfiles = array("sendfacrecmail@sendfacrecmail");
+		$this->langfiles = array("sendrecurringinvoicebymail@sendrecurringinvoicebymail");
 		//$this->phpmin = array(5,4);					// Minimum version of PHP required by module
-		$this->need_dolibarr_version = array(9,9);		// Minimum version of Dolibarr required by module
+		$this->need_dolibarr_version = array(10,0);		// Minimum version of Dolibarr required by module
 		$this->warnings_activation = array();			// Warning to show when we activate module. array('always'='text') or array('FR'='textfr','ES'='textes'...)
 		$this->warnings_activation_ext = array();		// Warning to show when we activate an external module. array('always'='text') or array('FR'='textfr','ES'='textes'...)
-		//$this->automatic_activation = array('FR'=>'sendfacrecmailWasAutomaticallyActivatedBecauseOfYourCountryChoice');
+		//$this->automatic_activation = array('FR'=>'sendrecurringinvoicebymailWasAutomaticallyActivatedBecauseOfYourCountryChoice');
 		//$this->always_enabled = true;								// If true, can't be disabled
 
 		// Constants
 		// List of particular constants to add when module is enabled (key, 'chaine', value, desc, visible, 'current' or 'allentities', deleteonunactive)
-		// Example: $this->const=array(0=>array('SENDFACRECMAIL_MYNEWCONST1','chaine','myvalue','This is a constant to add',1),
-		//                             1=>array('SENDFACRECMAIL_MYNEWCONST2','chaine','myvalue','This is another constant to add',0, 'current', 1)
+		// Example: $this->const=array(0=>array('SENDRECURRINGINVOICEBYMAIL_MYNEWCONST1','chaine','myvalue','This is a constant to add',1),
+		//                             1=>array('SENDRECURRINGINVOICEBYMAIL_MYNEWCONST2','chaine','myvalue','This is another constant to add',0, 'current', 1)
 		// );
 		$this->const = array(
-			//1=>array('SENDFACRECMAIL_MYCONSTANT', 'chaine', 'avalue', 'This is a constant to add', 1, 'allentities', 1)
+			//1=>array('SENDRECURRINGINVOICEBYMAIL_MYCONSTANT', 'chaine', 'avalue', 'This is a constant to add', 1, 'allentities', 1)
 		);
 
 		// Some keys to add into the overwriting translation tables
@@ -131,18 +131,18 @@ class modsendfacrecmail extends DolibarrModules
 			'fr_FR:ParentCompany'=>'Maison mère ou revendeur'
 		)*/
 
-		if (! isset($conf->sendfacrecmail) || ! isset($conf->sendfacrecmail->enabled))
+		if (! isset($conf->sendrecurringinvoicebymail) || ! isset($conf->sendrecurringinvoicebymail->enabled))
 		{
-			$conf->sendfacrecmail=new stdClass();
-			$conf->sendfacrecmail->enabled=0;
+			$conf->sendrecurringinvoicebymail=new stdClass();
+			$conf->sendrecurringinvoicebymail->enabled=0;
 		}
 
 
 		// Array to add new pages in new tabs
         $this->tabs = array();
 		// Example:
-		// $this->tabs[] = array('data'=>'objecttype:+tabname1:Title1:mylangfile@sendfacrecmail:$user->rights->sendfacrecmail->read:/sendfacrecmail/mynewtab1.php?id=__ID__');  					// To add a new tab identified by code tabname1
-        // $this->tabs[] = array('data'=>'objecttype:+tabname2:SUBSTITUTION_Title2:mylangfile@sendfacrecmail:$user->rights->othermodule->read:/sendfacrecmail/mynewtab2.php?id=__ID__',  	// To add another new tab identified by code tabname2. Label will be result of calling all substitution functions on 'Title2' key.
+		// $this->tabs[] = array('data'=>'objecttype:+tabname1:Title1:mylangfile@sendrecurringinvoicebymail:$user->rights->sendrecurringinvoicebymail->read:/sendrecurringinvoicebymail/mynewtab1.php?id=__ID__');  					// To add a new tab identified by code tabname1
+        // $this->tabs[] = array('data'=>'objecttype:+tabname2:SUBSTITUTION_Title2:mylangfile@sendrecurringinvoicebymail:$user->rights->othermodule->read:/sendrecurringinvoicebymail/mynewtab2.php?id=__ID__',  	// To add another new tab identified by code tabname2. Label will be result of calling all substitution functions on 'Title2' key.
         // $this->tabs[] = array('data'=>'objecttype:-tabname:NU:conditiontoremove');                                                     										// To remove an existing tab identified by code tabname
         //
         // Where objecttype can be
@@ -171,7 +171,7 @@ class modsendfacrecmail extends DolibarrModules
 		$this->dictionaries=array();
         /* Example:
         $this->dictionaries=array(
-            'langs'=>'mylangfile@sendfacrecmail',
+            'langs'=>'mylangfile@sendrecurringinvoicebymail',
             'tabname'=>array(MAIN_DB_PREFIX."table1",MAIN_DB_PREFIX."table2",MAIN_DB_PREFIX."table3"),		// List of tables we want to see into dictonnary editor
             'tablib'=>array("Table1","Table2","Table3"),													// Label of tables
             'tabsql'=>array('SELECT f.rowid as rowid, f.code, f.label, f.active FROM '.MAIN_DB_PREFIX.'table1 as f','SELECT f.rowid as rowid, f.code, f.label, f.active FROM '.MAIN_DB_PREFIX.'table2 as f','SELECT f.rowid as rowid, f.code, f.label, f.active FROM '.MAIN_DB_PREFIX.'table3 as f'),	// Request to select fields
@@ -180,27 +180,27 @@ class modsendfacrecmail extends DolibarrModules
             'tabfieldvalue'=>array("code,label","code,label","code,label"),																				// List of fields (list of fields to edit a record)
             'tabfieldinsert'=>array("code,label","code,label","code,label"),																			// List of fields (list of fields for insert)
             'tabrowid'=>array("rowid","rowid","rowid"),																									// Name of columns with primary key (try to always name it 'rowid')
-            'tabcond'=>array($conf->sendfacrecmail->enabled,$conf->sendfacrecmail->enabled,$conf->sendfacrecmail->enabled)												// Condition to show each dictionary
+            'tabcond'=>array($conf->sendrecurringinvoicebymail->enabled,$conf->sendrecurringinvoicebymail->enabled,$conf->sendrecurringinvoicebymail->enabled)												// Condition to show each dictionary
         );
         */
 
 
         // Boxes/Widgets
-		// Add here list of php file(s) stored in sendfacrecmail/core/boxes that contains class to show a widget.
+		// Add here list of php file(s) stored in sendrecurringinvoicebymail/core/boxes that contains class to show a widget.
         $this->boxes = array(
-        	//0=>array('file'=>'sendfacrecmailwidget1.php@sendfacrecmail','note'=>'Widget provided by sendfacrecmail','enabledbydefaulton'=>'Home'),
-        	//1=>array('file'=>'sendfacrecmailwidget2.php@sendfacrecmail','note'=>'Widget provided by sendfacrecmail'),
-        	//2=>array('file'=>'sendfacrecmailwidget3.php@sendfacrecmail','note'=>'Widget provided by sendfacrecmail')
+        	//0=>array('file'=>'sendrecurringinvoicebymailwidget1.php@sendrecurringinvoicebymail','note'=>'Widget provided by sendrecurringinvoicebymail','enabledbydefaulton'=>'Home'),
+        	//1=>array('file'=>'sendrecurringinvoicebymailwidget2.php@sendrecurringinvoicebymail','note'=>'Widget provided by sendrecurringinvoicebymail'),
+        	//2=>array('file'=>'sendrecurringinvoicebymailwidget3.php@sendrecurringinvoicebymail','note'=>'Widget provided by sendrecurringinvoicebymail')
         );
 
 
 		// Cronjobs (List of cron jobs entries to add when module is enabled)
 		// unit_frequency must be 60 for minute, 3600 for hour, 86400 for day, 604800 for week
 		$this->cronjobs = array(
-			//0=>array('label'=>'MyJob label', 'jobtype'=>'method', 'class'=>'/sendfacrecmail/class/myobject.class.php', 'objectname'=>'MyObject', 'method'=>'doScheduledJob', 'parameters'=>'', 'comment'=>'Comment', 'frequency'=>2, 'unitfrequency'=>3600, 'status'=>0, 'test'=>'$conf->sendfacrecmail->enabled', 'priority'=>50)
+			//0=>array('label'=>'MyJob label', 'jobtype'=>'method', 'class'=>'/sendrecurringinvoicebymail/class/myobject.class.php', 'objectname'=>'MyObject', 'method'=>'doScheduledJob', 'parameters'=>'', 'comment'=>'Comment', 'frequency'=>2, 'unitfrequency'=>3600, 'status'=>0, 'test'=>'$conf->sendrecurringinvoicebymail->enabled', 'priority'=>50)
 		);
-		// Example: $this->cronjobs=array(0=>array('label'=>'My label', 'jobtype'=>'method', 'class'=>'/dir/class/file.class.php', 'objectname'=>'MyClass', 'method'=>'myMethod', 'parameters'=>'param1, param2', 'comment'=>'Comment', 'frequency'=>2, 'unitfrequency'=>3600, 'status'=>0, 'test'=>'$conf->sendfacrecmail->enabled', 'priority'=>50),
-		//                                1=>array('label'=>'My label', 'jobtype'=>'command', 'command'=>'', 'parameters'=>'param1, param2', 'comment'=>'Comment', 'frequency'=>1, 'unitfrequency'=>3600*24, 'status'=>0, 'test'=>'$conf->sendfacrecmail->enabled', 'priority'=>50)
+		// Example: $this->cronjobs=array(0=>array('label'=>'My label', 'jobtype'=>'method', 'class'=>'/dir/class/file.class.php', 'objectname'=>'MyClass', 'method'=>'myMethod', 'parameters'=>'param1, param2', 'comment'=>'Comment', 'frequency'=>2, 'unitfrequency'=>3600, 'status'=>0, 'test'=>'$conf->sendrecurringinvoicebymail->enabled', 'priority'=>50),
+		//                                1=>array('label'=>'My label', 'jobtype'=>'command', 'command'=>'', 'parameters'=>'param1, param2', 'comment'=>'Comment', 'frequency'=>1, 'unitfrequency'=>3600*24, 'status'=>0, 'test'=>'$conf->sendrecurringinvoicebymail->enabled', 'priority'=>50)
 		// );
 
 
@@ -209,24 +209,24 @@ class modsendfacrecmail extends DolibarrModules
 		/*
 		$r=0;
 		$this->rights[$r][0] = $this->numero + $r;	// Permission id (must not be already used)
-		$this->rights[$r][1] = 'Read myobject of sendfacrecmail';	// Permission label
+		$this->rights[$r][1] = 'Read myobject of sendrecurringinvoicebymail';	// Permission label
 		$this->rights[$r][3] = 1; 					// Permission by default for new user (0/1)
-		$this->rights[$r][4] = 'read';				// In php code, permission will be checked by test if ($user->rights->sendfacrecmail->level1->level2)
-		$this->rights[$r][5] = '';				    // In php code, permission will be checked by test if ($user->rights->sendfacrecmail->level1->level2)
+		$this->rights[$r][4] = 'read';				// In php code, permission will be checked by test if ($user->rights->sendrecurringinvoicebymail->level1->level2)
+		$this->rights[$r][5] = '';				    // In php code, permission will be checked by test if ($user->rights->sendrecurringinvoicebymail->level1->level2)
 
 		$r++;
 		$this->rights[$r][0] = $this->numero + $r;	// Permission id (must not be already used)
-		$this->rights[$r][1] = 'Create/Update myobject of sendfacrecmail';	// Permission label
+		$this->rights[$r][1] = 'Create/Update myobject of sendrecurringinvoicebymail';	// Permission label
 		$this->rights[$r][3] = 1; 					// Permission by default for new user (0/1)
-		$this->rights[$r][4] = 'write';				// In php code, permission will be checked by test if ($user->rights->sendfacrecmail->level1->level2)
-		$this->rights[$r][5] = '';				    // In php code, permission will be checked by test if ($user->rights->sendfacrecmail->level1->level2)
+		$this->rights[$r][4] = 'write';				// In php code, permission will be checked by test if ($user->rights->sendrecurringinvoicebymail->level1->level2)
+		$this->rights[$r][5] = '';				    // In php code, permission will be checked by test if ($user->rights->sendrecurringinvoicebymail->level1->level2)
 
 		$r++;
 		$this->rights[$r][0] = $this->numero + $r;	// Permission id (must not be already used)
-		$this->rights[$r][1] = 'Delete myobject of sendfacrecmail';	// Permission label
+		$this->rights[$r][1] = 'Delete myobject of sendrecurringinvoicebymail';	// Permission label
 		$this->rights[$r][3] = 1; 					// Permission by default for new user (0/1)
-		$this->rights[$r][4] = 'delete';				// In php code, permission will be checked by test if ($user->rights->sendfacrecmail->level1->level2)
-		$this->rights[$r][5] = '';				    // In php code, permission will be checked by test if ($user->rights->sendfacrecmail->level1->level2)
+		$this->rights[$r][4] = 'delete';				// In php code, permission will be checked by test if ($user->rights->sendrecurringinvoicebymail->level1->level2)
+		$this->rights[$r][5] = '';				    // In php code, permission will be checked by test if ($user->rights->sendrecurringinvoicebymail->level1->level2)
 		*/
 
 
@@ -240,14 +240,14 @@ class modsendfacrecmail extends DolibarrModules
 		/*
 		$this->menu[$r++]=array('fk_menu'=>'',			                // '' if this is a top menu. For left menu, use 'fk_mainmenu=xxx' or 'fk_mainmenu=xxx,fk_leftmenu=yyy' where xxx is mainmenucode and yyy is a leftmenucode
 								'type'=>'top',			                // This is a Top menu entry
-								'titre'=>'sendfacrecmail',
-								'mainmenu'=>'sendfacrecmail',
+								'titre'=>'sendrecurringinvoicebymail',
+								'mainmenu'=>'sendrecurringinvoicebymail',
 								'leftmenu'=>'',
-								'url'=>'/sendfacrecmail/sendfacrecmailindex.php',
-								'langs'=>'sendfacrecmail@sendfacrecmail',	        // Lang file to use (without .lang) by module. File must be in langs/code_CODE/ directory.
+								'url'=>'/sendrecurringinvoicebymail/sendrecurringinvoicebymailindex.php',
+								'langs'=>'sendrecurringinvoicebymail@sendrecurringinvoicebymail',	        // Lang file to use (without .lang) by module. File must be in langs/code_CODE/ directory.
 								'position'=>1000+$r,
-								'enabled'=>'$conf->sendfacrecmail->enabled',	// Define condition to show or hide menu entry. Use '$conf->sendfacrecmail->enabled' if entry must be visible if module is enabled.
-								'perms'=>'1',			                // Use 'perms'=>'$user->rights->sendfacrecmail->level1->level2' if you want your menu with a permission rules
+								'enabled'=>'$conf->sendrecurringinvoicebymail->enabled',	// Define condition to show or hide menu entry. Use '$conf->sendrecurringinvoicebymail->enabled' if entry must be visible if module is enabled.
+								'perms'=>'1',			                // Use 'perms'=>'$user->rights->sendrecurringinvoicebymail->level1->level2' if you want your menu with a permission rules
 								'target'=>'',
 								'user'=>2);				                // 0=Menu for internal users, 1=external users, 2=both
 		*/
@@ -255,28 +255,28 @@ class modsendfacrecmail extends DolibarrModules
 		/* END MODULEBUILDER TOPMENU */
 
 		/* BEGIN MODULEBUILDER LEFTMENU MYOBJECT
-		$this->menu[$r++]=array(	'fk_menu'=>'fk_mainmenu=sendfacrecmail',	    // '' if this is a top menu. For left menu, use 'fk_mainmenu=xxx' or 'fk_mainmenu=xxx,fk_leftmenu=yyy' where xxx is mainmenucode and yyy is a leftmenucode
+		$this->menu[$r++]=array(	'fk_menu'=>'fk_mainmenu=sendrecurringinvoicebymail',	    // '' if this is a top menu. For left menu, use 'fk_mainmenu=xxx' or 'fk_mainmenu=xxx,fk_leftmenu=yyy' where xxx is mainmenucode and yyy is a leftmenucode
 								'type'=>'left',			                // This is a Left menu entry
 								'titre'=>'List MyObject',
-								'mainmenu'=>'sendfacrecmail',
-								'leftmenu'=>'sendfacrecmail_myobject_list',
-								'url'=>'/sendfacrecmail/myobject_list.php',
-								'langs'=>'sendfacrecmail@sendfacrecmail',	        // Lang file to use (without .lang) by module. File must be in langs/code_CODE/ directory.
+								'mainmenu'=>'sendrecurringinvoicebymail',
+								'leftmenu'=>'sendrecurringinvoicebymail_myobject_list',
+								'url'=>'/sendrecurringinvoicebymail/myobject_list.php',
+								'langs'=>'sendrecurringinvoicebymail@sendrecurringinvoicebymail',	        // Lang file to use (without .lang) by module. File must be in langs/code_CODE/ directory.
 								'position'=>1000+$r,
-								'enabled'=>'$conf->sendfacrecmail->enabled',  // Define condition to show or hide menu entry. Use '$conf->sendfacrecmail->enabled' if entry must be visible if module is enabled. Use '$leftmenu==\'system\'' to show if leftmenu system is selected.
-								'perms'=>'1',			                // Use 'perms'=>'$user->rights->sendfacrecmail->level1->level2' if you want your menu with a permission rules
+								'enabled'=>'$conf->sendrecurringinvoicebymail->enabled',  // Define condition to show or hide menu entry. Use '$conf->sendrecurringinvoicebymail->enabled' if entry must be visible if module is enabled. Use '$leftmenu==\'system\'' to show if leftmenu system is selected.
+								'perms'=>'1',			                // Use 'perms'=>'$user->rights->sendrecurringinvoicebymail->level1->level2' if you want your menu with a permission rules
 								'target'=>'',
 								'user'=>2);				                // 0=Menu for internal users, 1=external users, 2=both
-		$this->menu[$r++]=array(	'fk_menu'=>'fk_mainmenu=sendfacrecmail,fk_leftmenu=sendfacrecmail',	    // '' if this is a top menu. For left menu, use 'fk_mainmenu=xxx' or 'fk_mainmenu=xxx,fk_leftmenu=yyy' where xxx is mainmenucode and yyy is a leftmenucode
+		$this->menu[$r++]=array(	'fk_menu'=>'fk_mainmenu=sendrecurringinvoicebymail,fk_leftmenu=sendrecurringinvoicebymail',	    // '' if this is a top menu. For left menu, use 'fk_mainmenu=xxx' or 'fk_mainmenu=xxx,fk_leftmenu=yyy' where xxx is mainmenucode and yyy is a leftmenucode
 								'type'=>'left',			                // This is a Left menu entry
 								'titre'=>'New MyObject',
-								'mainmenu'=>'sendfacrecmail',
-								'leftmenu'=>'sendfacrecmail_myobject_new',
-								'url'=>'/sendfacrecmail/myobject_page.php?action=create',
-								'langs'=>'sendfacrecmail@sendfacrecmail',	        // Lang file to use (without .lang) by module. File must be in langs/code_CODE/ directory.
+								'mainmenu'=>'sendrecurringinvoicebymail',
+								'leftmenu'=>'sendrecurringinvoicebymail_myobject_new',
+								'url'=>'/sendrecurringinvoicebymail/myobject_page.php?action=create',
+								'langs'=>'sendrecurringinvoicebymail@sendrecurringinvoicebymail',	        // Lang file to use (without .lang) by module. File must be in langs/code_CODE/ directory.
 								'position'=>1000+$r,
-								'enabled'=>'$conf->sendfacrecmail->enabled',  // Define condition to show or hide menu entry. Use '$conf->sendfacrecmail->enabled' if entry must be visible if module is enabled. Use '$leftmenu==\'system\'' to show if leftmenu system is selected.
-								'perms'=>'1',			                // Use 'perms'=>'$user->rights->sendfacrecmail->level1->level2' if you want your menu with a permission rules
+								'enabled'=>'$conf->sendrecurringinvoicebymail->enabled',  // Define condition to show or hide menu entry. Use '$conf->sendrecurringinvoicebymail->enabled' if entry must be visible if module is enabled. Use '$leftmenu==\'system\'' to show if leftmenu system is selected.
+								'perms'=>'1',			                // Use 'perms'=>'$user->rights->sendrecurringinvoicebymail->level1->level2' if you want your menu with a permission rules
 								'target'=>'',
 								'user'=>2);				                // 0=Menu for internal users, 1=external users, 2=both
 		END MODULEBUILDER LEFTMENU MYOBJECT */
@@ -287,10 +287,10 @@ class modsendfacrecmail extends DolibarrModules
 
 		/* BEGIN MODULEBUILDER EXPORT MYOBJECT */
 		/*
-		$langs->load("sendfacrecmail@sendfacrecmail");
+		$langs->load("sendrecurringinvoicebymail@sendrecurringinvoicebymail");
 		$this->export_code[$r]=$this->rights_class.'_'.$r;
 		$this->export_label[$r]='MyObjectLines';	// Translation key (used only if key ExportDataset_xxx_z not found)
-		$this->export_icon[$r]='myobject@sendfacrecmail';
+		$this->export_icon[$r]='myobject@sendrecurringinvoicebymail';
 		$keyforclass = 'MyObject'; $keyforclassfile='/mymobule/class/myobject.class.php'; $keyforelement='myobject';
 		include DOL_DOCUMENT_ROOT.'/core/commonfieldsinexport.inc.php';
 		$keyforselect='myobject'; $keyforaliasextra='extra'; $keyforelement='myobject';
@@ -317,11 +317,11 @@ class modsendfacrecmail extends DolibarrModules
 		$sql = array();
 
 		// we check if our model already exists
-		$result = $this->db->query("SELECT COUNT(*) AS cpt FROM " . MAIN_DB_PREFIX."c_email_templates WHERE module = 'sendfacrecmail'");
+		$result = $this->db->query("SELECT COUNT(*) AS cpt FROM " . MAIN_DB_PREFIX."c_email_templates WHERE module = 'sendrecurringinvoicebymail'");
 		if ($result) {
 			$row = $this->db->fetch_object($result);
 			if ($row->cpt == 0) {
-				$sql[] = "INSERT INTO " . MAIN_DB_PREFIX."c_email_templates (module, type_template, lang, label, joinfiles, topic, content) VALUES ('sendfacrecmail', 'thirdparty', 'fr_FR', 'Envoi automatique de facture via sendfacrecmail', '1', '[__MYCOMPANY_NAME__] Nouvelle facture __REF__', 'Bonjour,\n\nNouvelle facture ci-jointe.\n\nEn cas de règlement par virement, merci d''indiquer dans le libellé la référence du contrat : __CONTRACT_REF__ .\n\nCordialement,\nle nouveau robot de l''équipe tréso.')";
+				$sql[] = "INSERT INTO " . MAIN_DB_PREFIX."c_email_templates (module, type_template, lang, label, joinfiles, topic, content) VALUES ('sendrecurringinvoicebymail', 'thirdparty', 'fr_FR', 'Envoi automatique de facture via sendrecurringinvoicebymail', '1', '[__MYCOMPANY_NAME__] Nouvelle facture __REF__', 'Bonjour,\n\nNouvelle facture ci-jointe.\n\nEn cas de règlement par virement, merci d''indiquer dans le libellé la référence du contrat : __CONTRACT_REF__ .\n\nCordialement,\nle nouveau robot de l''équipe tréso.')";
 			}
 		}
 
@@ -340,7 +340,7 @@ class modsendfacrecmail extends DolibarrModules
 	{
 		$sql = array();
 
-		$sql[] = "DELETE FROM " . MAIN_DB_PREFIX."c_email_templates WHERE module = 'sendfacrecmail'";
+		$sql[] = "DELETE FROM " . MAIN_DB_PREFIX."c_email_templates WHERE module = 'sendrecurringinvoicebymail'";
 
 		return $this->_remove($sql, $options);
 	}