+++ ChangeLog.md
@@ -1,4 +1,8 @@
-# CHANGELOG SENDFACRECMAIL FOR <a href="https://www.dolibarr.org">DOLIBARR ERP CRM</a>
+# CHANGELOG SENDRECURRINGINVOICEBYMAIL FOR <a href="https://www.dolibarr.org">DOLIBARR ERP CRM</a>
+
+## 0.2.3
+Renaming from 'sendfacrecmail' to 'sendrecurringinvoicebymail'.
+
 
 ## 0.1.2
-Première version fonctionnelle.
+First working version.